Rename lib->integration to avoid a .gitignore name
